requirements.txt
# Core dependencies
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Data collection
python-binance>=1.0.19
pyarrow>=12.0.0  # For parquet support

# Feature engineering
ta-lib>=0.4.28  # Optional: for additional technical indicators

# Training & optimization
optuna>=3.3.0  # For hyperparameter optimization
ray[tune]>=2.7.0  # Alternative for distributed tuning

# Visualization (optional)
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Development
pytest>=7.4.0
black>=23.7.0
flake8>=6.1.0

